Added validation for billpay payments
 - changed billpay model to include a status field
 - payments won't reduce balances to below min

+++ Assignment 2/Background Services/BillPayBackgroundService.cs
@@ -6,6 +6,7 @@ using Microsoft.Extensions.Hosting;
 using System;
 using System.Threading;
 using System.Threading.Tasks;
+using DataValidator;
 
 namespace Assignment_2.Background_Services
 {
@@ -33,42 +34,51 @@ namespace Assignment_2.Background_Services
             var billpays = await context.BillPays.ToListAsync(stoppingToken);
             foreach (var bill in billpays)
             {
-                if (DateTime.Compare(bill.ScheduleTimeUtc, DateTime.Now) < 0)
+                if (DateTime.Compare(bill.ScheduleTimeUtc, DateTime.Now) < 0 && bill.Status.Equals(BillPayStatus.Ready))
                 {
                     var account = await context.Accounts.FindAsync(bill.AccountNumber);
-                    account.Balance -= bill.Amount;
-                    account.Transactions.Add(
-                    new Transaction
+                    if (account.Balance - bill.Amount < AccountChecks.GetAccountTypeMin(account.AccountType.ToString()))
                     {
-                        TransactionType = TransactionType.B,
-                        Amount = bill.Amount,
-                        TransactionTimeUtc = DateTime.UtcNow
-                    });
+                        bill.Status = BillPayStatus.Error;
+                        context.BillPays.Update(bill);
+                    }
+                    else
+                    {
 
-                }
-                switch (bill.Period)
-                {
-                    case PaymentPeriod.O:
-                        context.BillPays.Remove(bill);
-                        break;
-                    case PaymentPeriod.M:
-                        bill.ScheduleTimeUtc = bill.ScheduleTimeUtc.AddMonths(1);
-                        context.BillPays.Update(bill);
-                        break;
-                    case PaymentPeriod.Q:
-                        bill.ScheduleTimeUtc = bill.ScheduleTimeUtc.AddMonths(3);
-                        context.BillPays.Update(bill);
-                        break;
-                    case PaymentPeriod.Y:
-                        bill.ScheduleTimeUtc = bill.ScheduleTimeUtc.AddYears(1);
-                        context.BillPays.Update(bill);
-                        break;
+                        account.Balance -= bill.Amount;
+                        account.Transactions.Add(
+                        new Transaction
+                        {
+                            TransactionType = TransactionType.B,
+                            Amount = bill.Amount,
+                            TransactionTimeUtc = DateTime.UtcNow
+                        });
 
+
+                        switch (bill.Period)
+                        {
+                            case PaymentPeriod.O:
+                                context.BillPays.Remove(bill);
+                                break;
+                            case PaymentPeriod.M:
+                                bill.ScheduleTimeUtc = bill.ScheduleTimeUtc.AddMonths(1);
+                                context.BillPays.Update(bill);
+                                break;
+                            case PaymentPeriod.Q:
+                                bill.ScheduleTimeUtc = bill.ScheduleTimeUtc.AddMonths(3);
+                                context.BillPays.Update(bill);
+                                break;
+                            case PaymentPeriod.Y:
+                                bill.ScheduleTimeUtc = bill.ScheduleTimeUtc.AddYears(1);
+                                context.BillPays.Update(bill);
+                                break;
+
+                        }
+                    }
                 }
+
             }
-
             await context.SaveChangesAsync(stoppingToken);
-
         }
     }
 }
